Behavior-preserving refactor; `List`, `Bag`, and `Tuple` in own files
