+++ partiql-value/src/lib.rs
@@ -1,19 +1,24 @@
-use itertools::Itertools;
 use ordered_float::OrderedFloat;
 use std::cmp::Ordering;
 
 use std::borrow::Cow;
-use std::collections::HashSet;
+
 use std::fmt::{Debug, Formatter};
-use std::hash::{Hash, Hasher};
-use std::iter::zip;
-use std::{ops, slice, vec};
+use std::hash::Hash;
+
+use std::{ops, vec};
 
 use rust_decimal::prelude::FromPrimitive;
 use rust_decimal::{Decimal as RustDecimal, Decimal};
-use unicase::UniCase;
 
+mod bag;
 pub mod ion;
+mod list;
+mod tuple;
+
+pub use bag::*;
+pub use list::*;
+pub use tuple::*;
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -543,7 +548,7 @@ impl Value {
         if let Value::Bag(b) = self {
             *b
         } else {
-            Bag(vec![self])
+            Bag::from(vec![self])
         }
     }
 
@@ -561,7 +566,7 @@ impl Value {
         if let Value::List(b) = self {
             *b
         } else {
-            List(vec![self])
+            List::from(vec![self])
         }
     }
 
@@ -855,566 +860,6 @@ impl From<Bag> for Value {
     fn from(v: Bag) -> Self {
         Value::Bag(Box::new(v))
     }
-}
-
-#[derive(Default, Hash, PartialEq, Eq, Clone)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-/// Represents a PartiQL List value, e.g. [1, 2, 'one']
-pub struct List(Vec<Value>);
-
-impl List {
-    #[inline]
-    pub fn push(&mut self, value: Value) {
-        self.0.push(value);
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    #[inline]
-    pub fn get(&self, idx: i64) -> Option<&Value> {
-        self.0.get(idx as usize)
-    }
-
-    #[inline]
-    pub fn get_mut(&mut self, idx: i64) -> Option<&mut Value> {
-        self.0.get_mut(idx as usize)
-    }
-
-    #[inline]
-    pub fn iter(&self) -> ListIter {
-        ListIter(self.0.iter())
-    }
-
-    #[inline]
-    pub fn reserve(&mut self, additional: usize) {
-        self.0.reserve(additional);
-    }
-}
-
-impl Extend<Value> for List {
-    #[inline]
-    fn extend<Iter: IntoIterator<Item = Value>>(&mut self, iter: Iter) {
-        let iterator = iter.into_iter();
-        let (lower_bound, _) = iterator.size_hint();
-        self.reserve(lower_bound);
-        iterator.for_each(move |v| self.push(v));
-    }
-}
-
-impl From<Vec<Value>> for List {
-    #[inline]
-    fn from(values: Vec<Value>) -> Self {
-        List(values)
-    }
-}
-
-impl From<Bag> for List {
-    #[inline]
-    fn from(bag: Bag) -> Self {
-        List(bag.0)
-    }
-}
-
-impl<T> FromIterator<T> for List
-where
-    T: Into<Value>,
-{
-    #[inline]
-    fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> List {
-        let iterator = iter.into_iter().map(Into::into);
-        iterator.collect::<Vec<_>>().into()
-    }
-}
-
-#[macro_export]
-macro_rules! partiql_list {
-    () => (
-         List::from(vec![])
-    );
-    ($elem:expr; $n:expr) => (
-        List::from(vec![Value::from($elem); $n])
-    );
-    ($($x:expr),+ $(,)?) => (
-        List::from(vec![$(Value::from($x)),+])
-    );
-}
-
-#[derive(Debug, Clone)]
-pub struct ListIter<'a>(slice::Iter<'a, Value>);
-
-impl<'a> Iterator for ListIter<'a> {
-    type Item = &'a Value;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-}
-
-impl IntoIterator for List {
-    type Item = Value;
-    type IntoIter = ListIntoIterator;
-
-    fn into_iter(self) -> ListIntoIterator {
-        ListIntoIterator(self.0.into_iter())
-    }
-}
-
-pub struct ListIntoIterator(vec::IntoIter<Value>);
-
-impl Iterator for ListIntoIterator {
-    type Item = Value;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-}
-
-impl Debug for List {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_list().entries(&self.0).finish()
-    }
-}
-
-impl PartialOrd for List {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        let mut l = self.0.iter();
-        let mut r = other.0.iter();
-
-        loop {
-            match (l.next(), r.next()) {
-                (None, None) => return Some(Ordering::Equal),
-                (Some(_), None) => return Some(Ordering::Greater),
-                (None, Some(_)) => return Some(Ordering::Less),
-                (Some(lv), Some(rv)) => match lv.partial_cmp(rv) {
-                    None => return None,
-                    Some(Ordering::Less) => return Some(Ordering::Less),
-                    Some(Ordering::Greater) => return Some(Ordering::Greater),
-                    Some(Ordering::Equal) => continue,
-                },
-            }
-        }
-    }
-}
-
-impl Ord for List {
-    fn cmp(&self, other: &Self) -> Ordering {
-        let mut l = self.0.iter();
-        let mut r = other.0.iter();
-
-        loop {
-            match (l.next(), r.next()) {
-                (None, None) => return Ordering::Equal,
-                (Some(_), None) => return Ordering::Greater,
-                (None, Some(_)) => return Ordering::Less,
-                (Some(lv), Some(rv)) => match lv.cmp(rv) {
-                    Ordering::Less => return Ordering::Less,
-                    Ordering::Greater => return Ordering::Greater,
-                    Ordering::Equal => continue,
-                },
-            }
-        }
-    }
-}
-
-#[derive(Default, Eq, Clone)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-/// Represents a PartiQL BAG value, e.g.: <<1, 'two', 4>>
-pub struct Bag(Vec<Value>);
-
-impl Bag {
-    #[inline]
-    pub fn push(&mut self, value: Value) {
-        self.0.push(value);
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    #[inline]
-    pub fn iter(&self) -> BagIter {
-        BagIter(self.0.iter())
-    }
-
-    #[inline]
-    pub fn reserve(&mut self, additional: usize) {
-        self.0.reserve(additional);
-    }
-}
-
-impl Extend<Value> for Bag {
-    #[inline]
-    fn extend<Iter: IntoIterator<Item = Value>>(&mut self, iter: Iter) {
-        let iterator = iter.into_iter();
-        let (lower_bound, _) = iterator.size_hint();
-        self.reserve(lower_bound);
-        iterator.for_each(move |v| self.push(v));
-    }
-}
-
-impl From<Vec<Value>> for Bag {
-    #[inline]
-    fn from(values: Vec<Value>) -> Self {
-        Bag(values)
-    }
-}
-
-impl From<HashSet<Value>> for Bag {
-    #[inline]
-    fn from(values: HashSet<Value>) -> Self {
-        Bag(values.into_iter().collect())
-    }
-}
-
-impl From<List> for Bag {
-    #[inline]
-    fn from(list: List) -> Self {
-        Bag(list.0)
-    }
-}
-
-impl<T> FromIterator<T> for Bag
-where
-    T: Into<Value>,
-{
-    #[inline]
-    fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> Bag {
-        let iterator = iter.into_iter().map(Into::into);
-        iterator.collect::<Vec<_>>().into()
-    }
-}
-
-#[macro_export]
-macro_rules! partiql_bag {
-    () => (
-         Bag::from(vec![])
-    );
-    ($elem:expr; $n:expr) => (
-        Bag::from(vec![Value::from($elem); $n])
-    );
-    ($($x:expr),+ $(,)?) => (
-        Bag::from(vec![$(Value::from($x)),+])
-    );
-}
-
-#[derive(Debug, Clone)]
-pub struct BagIter<'a>(slice::Iter<'a, Value>);
-
-impl<'a> Iterator for BagIter<'a> {
-    type Item = &'a Value;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-}
-
-impl IntoIterator for Bag {
-    type Item = Value;
-    type IntoIter = BagIntoIterator;
-
-    fn into_iter(self) -> BagIntoIterator {
-        BagIntoIterator(self.0.into_iter())
-    }
-}
-
-pub struct BagIntoIterator(vec::IntoIter<Value>);
-
-impl Iterator for BagIntoIterator {
-    type Item = Value;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-}
-
-impl Debug for Bag {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "<<")?;
-        let mut iter = self.iter().peekable();
-        while let Some(v) = iter.next() {
-            if iter.peek().is_some() {
-                write!(f, "{v:?}, ")?;
-            } else {
-                write!(f, "{v:?}")?;
-            }
-        }
-        write!(f, ">>")
-    }
-}
-
-impl PartialEq for Bag {
-    fn eq(&self, other: &Self) -> bool {
-        if self.len() != other.len() {
-            return false;
-        }
-
-        let lhs = self.0.iter().sorted();
-        let rhs = other.0.iter().sorted();
-        lhs.zip(rhs).all(|(l, r)| l == r)
-    }
-}
-
-impl PartialOrd for Bag {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for Bag {
-    fn cmp(&self, other: &Self) -> Ordering {
-        let mut l = self.0.clone();
-        l.sort();
-        let mut r = other.0.clone();
-        r.sort();
-        List(l).cmp(&List(r))
-    }
-}
-
-impl Hash for Bag {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        for v in self.0.iter().sorted() {
-            v.hash(state);
-        }
-    }
-}
-
-#[derive(Default, Eq, Clone)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct Tuple {
-    attrs: Vec<String>,
-    vals: Vec<Value>,
-}
-
-impl Tuple {
-    pub fn new() -> Self {
-        Tuple {
-            attrs: vec![],
-            vals: vec![],
-        }
-    }
-
-    #[inline]
-    pub fn insert(&mut self, attr: &str, val: Value) {
-        self.attrs.push(attr.to_string());
-        self.vals.push(val);
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.attrs.len()
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    #[inline]
-    /// Creates a `Tuple` containing all the attributes and values provided by `self` and `other`
-    /// removing duplicate attributes. Assumes that `self` contains unique attributes and `other`
-    /// contains unique attributes. In the case of duplicate attributes between `self` and `other`,
-    /// the result `Tuple` will contain the attribute provided by `other`. See section 3.4 of the
-    /// spec for details: https://partiql.org/assets/PartiQL-Specification.pdf#subsection.3.4.
-    pub fn tuple_concat(&self, other: &Tuple) -> Self {
-        other
-            .pairs()
-            .chain(self.pairs())
-            .map(|(a, v)| (a, v.clone()))
-            .unique_by(|(a, _)| *a)
-            .collect()
-    }
-
-    #[inline]
-    pub fn get(&self, attr: &BindingsName) -> Option<&Value> {
-        match attr {
-            BindingsName::CaseSensitive(s) => match self.attrs.iter().position(|a| a.as_str() == s)
-            {
-                Some(i) => Some(&self.vals[i]),
-                _ => None,
-            },
-            BindingsName::CaseInsensitive(s) => match self
-                .attrs
-                .iter()
-                .position(|a| UniCase::<&String>::from(a) == UniCase::<&String>::from(s))
-            {
-                Some(i) => Some(&self.vals[i]),
-                _ => None,
-            },
-        }
-    }
-
-    #[inline]
-    pub fn remove(&mut self, attr: &BindingsName) -> Option<Value> {
-        match attr {
-            BindingsName::CaseSensitive(s) => match self.attrs.iter().position(|a| a.as_str() == s)
-            {
-                Some(i) => {
-                    self.attrs.remove(i);
-                    Some(self.vals.remove(i))
-                }
-                _ => None,
-            },
-            BindingsName::CaseInsensitive(s) => match self
-                .attrs
-                .iter()
-                .position(|a| UniCase::<&String>::from(a) == UniCase::<&String>::from(s))
-            {
-                Some(i) => {
-                    self.attrs.remove(i);
-                    Some(self.vals.remove(i))
-                }
-                _ => None,
-            },
-        }
-    }
-
-    #[inline]
-    pub fn pairs(&self) -> impl Iterator<Item = (&str, &Value)> + Clone {
-        zip(&self.attrs, &self.vals).map(|(k, v)| (k.as_str(), v))
-    }
-
-    #[inline]
-    pub fn into_pairs(self) -> impl Iterator<Item = (String, Value)> {
-        zip(self.attrs, self.vals)
-    }
-
-    #[inline]
-    pub fn values(&self) -> impl Iterator<Item = &Value> + Clone {
-        self.vals.iter()
-    }
-
-    #[inline]
-    pub fn into_values(self) -> impl Iterator<Item = Value> {
-        self.vals.into_iter()
-    }
-}
-
-impl<const N: usize, T> From<[(&str, T); N]> for Tuple
-where
-    T: Into<Value>,
-{
-    #[inline]
-    fn from(arr: [(&str, T); N]) -> Self {
-        arr.into_iter()
-            .fold(Tuple::new(), |mut acc: Tuple, (attr, val)| {
-                acc.insert(attr, val.into());
-                acc
-            })
-    }
-}
-
-impl<S, T> FromIterator<(S, T)> for Tuple
-where
-    S: Into<String>,
-    T: Into<Value>,
-{
-    #[inline]
-    fn from_iter<I: IntoIterator<Item = (S, T)>>(iter: I) -> Tuple {
-        let iterator = iter.into_iter();
-        let (lower, _) = iterator.size_hint();
-        let mut attrs = Vec::with_capacity(lower);
-        let mut vals = Vec::with_capacity(lower);
-        for (k, v) in iterator {
-            attrs.push(k.into());
-            vals.push(v.into());
-        }
-        Tuple { attrs, vals }
-    }
-}
-
-impl Iterator for Tuple {
-    type Item = (String, Value);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match (self.attrs.pop(), self.vals.pop()) {
-            (Some(attr), Some(val)) => Some((attr, val)),
-            _ => None,
-        }
-    }
-}
-
-impl PartialEq for Tuple {
-    fn eq(&self, other: &Self) -> bool {
-        let s1: HashSet<(&str, &Value)> = self.pairs().collect();
-        let s2: HashSet<(&str, &Value)> = other.pairs().collect();
-        s1.eq(&s2)
-    }
-}
-
-impl PartialOrd for Tuple {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Hash for Tuple {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        for (k, v) in self.pairs() {
-            k.hash(state);
-            v.hash(state);
-        }
-    }
-}
-
-impl Debug for Tuple {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{{")?;
-        let mut iter = self.pairs().peekable();
-        while let Some((k, v)) = iter.next() {
-            if iter.peek().is_some() {
-                write!(f, " {k}: {v:?},")?;
-            } else {
-                write!(f, " {k}: {v:?} ")?;
-            }
-        }
-        write!(f, "}}")
-    }
-}
-
-impl Ord for Tuple {
-    fn cmp(&self, other: &Self) -> Ordering {
-        let self_pairs = self.pairs();
-        let other_pairs = other.pairs();
-        let mut p1 = self_pairs.sorted();
-        let mut p2 = other_pairs.sorted();
-
-        loop {
-            return match (p1.next(), p2.next()) {
-                (None, None) => Ordering::Equal,
-                (Some(_), None) => Ordering::Greater,
-                (None, Some(_)) => Ordering::Less,
-                (Some(lv), Some(rv)) => match lv.cmp(&rv) {
-                    Ordering::Less => Ordering::Less,
-                    Ordering::Greater => Ordering::Greater,
-                    Ordering::Equal => continue,
-                },
-            };
-        }
-    }
-}
-
-#[macro_export]
-macro_rules! partiql_tuple {
-    () => (
-         Tuple::new()
-    );
-    ($(($x:expr, $y:expr)),+ $(,)?) => (
-        Tuple::from([$(($x, Value::from($y))),+])
-    );
 }
 
 #[cfg(test)]
