+++ partiql-value/src/list.rs
@@ -1,0 +1,182 @@
+use std::cmp::Ordering;
+
+use std::fmt::{Debug, Formatter};
+use std::hash::Hash;
+
+use std::{slice, vec};
+
+use crate::{Bag, Value};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[derive(Default, Hash, PartialEq, Eq, Clone)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+/// Represents a PartiQL List value, e.g. [1, 2, 'one']
+pub struct List(Vec<Value>);
+
+impl List {
+    #[inline]
+    pub fn push(&mut self, value: Value) {
+        self.0.push(value);
+    }
+
+    #[inline]
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    #[inline]
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    #[inline]
+    pub fn get(&self, idx: i64) -> Option<&Value> {
+        self.0.get(idx as usize)
+    }
+
+    #[inline]
+    pub fn get_mut(&mut self, idx: i64) -> Option<&mut Value> {
+        self.0.get_mut(idx as usize)
+    }
+
+    #[inline]
+    pub fn iter(&self) -> ListIter {
+        ListIter(self.0.iter())
+    }
+
+    #[inline]
+    pub fn reserve(&mut self, additional: usize) {
+        self.0.reserve(additional);
+    }
+
+    #[inline]
+    pub(crate) fn values(self) -> Vec<Value> {
+        self.0
+    }
+}
+
+impl Extend<Value> for List {
+    #[inline]
+    fn extend<Iter: IntoIterator<Item = Value>>(&mut self, iter: Iter) {
+        let iterator = iter.into_iter();
+        let (lower_bound, _) = iterator.size_hint();
+        self.reserve(lower_bound);
+        iterator.for_each(move |v| self.push(v));
+    }
+}
+
+impl From<Vec<Value>> for List {
+    #[inline]
+    fn from(values: Vec<Value>) -> Self {
+        List(values)
+    }
+}
+
+impl From<Bag> for List {
+    #[inline]
+    fn from(bag: Bag) -> Self {
+        List(bag.values())
+    }
+}
+
+impl<T> FromIterator<T> for List
+where
+    T: Into<Value>,
+{
+    #[inline]
+    fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> List {
+        let iterator = iter.into_iter().map(Into::into);
+        iterator.collect::<Vec<_>>().into()
+    }
+}
+
+#[macro_export]
+macro_rules! partiql_list {
+    () => (
+         List::from(vec![])
+    );
+    ($elem:expr; $n:expr) => (
+        List::from(vec![Value::from($elem); $n])
+    );
+    ($($x:expr),+ $(,)?) => (
+        List::from(vec![$(Value::from($x)),+])
+    );
+}
+
+#[derive(Debug, Clone)]
+pub struct ListIter<'a>(slice::Iter<'a, Value>);
+
+impl<'a> Iterator for ListIter<'a> {
+    type Item = &'a Value;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next()
+    }
+}
+
+impl IntoIterator for List {
+    type Item = Value;
+    type IntoIter = ListIntoIterator;
+
+    fn into_iter(self) -> ListIntoIterator {
+        ListIntoIterator(self.0.into_iter())
+    }
+}
+
+pub struct ListIntoIterator(vec::IntoIter<Value>);
+
+impl Iterator for ListIntoIterator {
+    type Item = Value;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next()
+    }
+}
+
+impl Debug for List {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.debug_list().entries(&self.0).finish()
+    }
+}
+
+impl PartialOrd for List {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        let mut l = self.0.iter();
+        let mut r = other.0.iter();
+
+        loop {
+            match (l.next(), r.next()) {
+                (None, None) => return Some(Ordering::Equal),
+                (Some(_), None) => return Some(Ordering::Greater),
+                (None, Some(_)) => return Some(Ordering::Less),
+                (Some(lv), Some(rv)) => match lv.partial_cmp(rv) {
+                    None => return None,
+                    Some(Ordering::Less) => return Some(Ordering::Less),
+                    Some(Ordering::Greater) => return Some(Ordering::Greater),
+                    Some(Ordering::Equal) => continue,
+                },
+            }
+        }
+    }
+}
+
+impl Ord for List {
+    fn cmp(&self, other: &Self) -> Ordering {
+        let mut l = self.0.iter();
+        let mut r = other.0.iter();
+
+        loop {
+            match (l.next(), r.next()) {
+                (None, None) => return Ordering::Equal,
+                (Some(_), None) => return Ordering::Greater,
+                (None, Some(_)) => return Ordering::Less,
+                (Some(lv), Some(rv)) => match lv.cmp(rv) {
+                    Ordering::Less => return Ordering::Less,
+                    Ordering::Greater => return Ordering::Greater,
+                    Ordering::Equal => continue,
+                },
+            }
+        }
+    }
+}
